Extract core DTO classes into separate files
