+++ DataModels/Core/Tweet.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Twity.DataModels.Core
+{
+    [Serializable]
+    public class Tweet : TweetObjectWithUser
+    {
+        public TweetObjectWithUser retweeted_status;
+    }
+}
