+++ DataModels/Core/TweetObjectWithUser.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Twity.DataModels.Core
+{
+    [Serializable]
+    public class TweetObjectWithUser : TweetObject
+    {
+        public TweetUser user;
+    }
+}
