+++ DataModels/Core/UploadMediaImage.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Twity.DataModels.Core
+{
+    [Serializable]
+    public class UploadMediaImage
+    {
+        public string image_type;
+        public int w;
+        public int h;
+    }
+}
