+++ Helpers/JsonHelper.cs
@@ -1,0 +1,15 @@
+﻿
+namespace Twity.Helpers
+{
+    public class JsonHelper {
+
+        public static string ArrayToObject (string arrayString) {
+            if (arrayString.StartsWith ("[")) {
+                arrayString = "{\"items\":" + arrayString + "}";
+                return arrayString;
+            } else {
+                return arrayString;
+            }
+        }
+    }
+}
