--- scripts/TwitterHelper.cs
@@ -1,91 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Text.RegularExpressions;
-
-namespace Twity
-{
-    public class Helper
-    {
-
-        public static SortedDictionary<string, string> ConvertToSortedDictionary(Dictionary<string, string> APIParams)
-        {
-            SortedDictionary<string, string> parameters = new SortedDictionary<string, string>();
-            foreach (KeyValuePair<string, string> APIParam in APIParams)
-            {
-                parameters.Add(APIParam.Key, APIParam.Value);
-            }
-            return parameters;
-        }
-
-        public static string GenerateRequestparams(SortedDictionary<string, string> parameters)
-        {
-            StringBuilder requestParams = new StringBuilder();
-            foreach (KeyValuePair<string, string> param in parameters)
-            {
-                requestParams.Append(Helper.UrlEncode(param.Key) + "=" + Helper.UrlEncode(param.Value) + "&");
-            }
-            requestParams.Length -= 1; // Remove "&" at the last of string
-            return requestParams.ToString();
-        }
-
-        // The below helper method is modified from "WebRequestBuilder.cs" in Twitterizer(http://www.twitterizer.net/).
-        // Here is its license.
-
-        //-----------------------------------------------------------------------
-        // <copyright file="WebRequestBuilder.cs" company="Patrick 'Ricky' Smith">
-        //  This file is part of the Twitterizer library (http://www.twitterizer.net/)
-        //
-        //  Copyright (c) 2010, Patrick "Ricky" Smith (ricky@digitally-born.com)
-        //  All rights reserved.
-        //
-        //  Redistribution and use in source and binary forms, with or without modification, are
-        //  permitted provided that the following conditions are met:
-        //
-        //  - Redistributions of source code must retain the above copyright notice, this list
-        //    of conditions and the following disclaimer.
-        //  - Redistributions in binary form must reproduce the above copyright notice, this list
-        //    of conditions and the following disclaimer in the documentation and/or other
-        //    materials provided with the distribution.
-        //  - Neither the name of the Twitterizer nor the names of its contributors may be
-        //    used to endorse or promote products derived from this software without specific
-        //    prior written permission.
-        //
-        //  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-        //  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-        //  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
-        //  IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
-        //  INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
-        //  NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-        //  PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-        //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-        //  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-        //  POSSIBILITY OF SUCH DAMAGE.
-        // </copyright>
-        // <author>Ricky Smith</author>
-        // <summary>Provides the means of preparing and executing Anonymous and OAuth signed web requests.</summary>
-        //-----------------------------------------------------------------------
-
-        public static string UrlEncode(string value)
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                return string.Empty;
-            }
-
-            value = Uri.EscapeDataString(value);
-            value = Regex.Replace(value, "(%[0-9a-f][0-9a-f])", c => c.Value.ToUpper());
-            value = value
-                .Replace("(", "%28")
-                .Replace(")", "%29")
-                .Replace("$", "%24")
-                .Replace("!", "%21")
-                .Replace("*", "%2A")
-                .Replace("'", "%27");
-            value = value.Replace("%7E", "~");
-            return value;
-
-        }
-    }
-
-}
