+++ scripts/TwitterOauth.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Security.Cryptography;
 using System.Text;
+using Twity.Helpers;
 
 namespace Twity
 {
