+++ scripts/TwitterStream.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.Text;
+using Twity.DataModels.Core;
+using Twity.Helpers;
 using UnityEngine;
 using UnityEngine.Networking;
