+++ scripts/twitterJson.cs
@@ -1,117 +1,7 @@
 ﻿using System;
+using Twity.DataModels.Core;
 
 namespace Twity {
-
-    #region Basic Class
-    [Serializable]
-    public class TweetObject
-    {
-        public long id;
-        public string id_str;
-        public string created_at;
-        public string text;
-
-        public string in_reply_to_screen_name;
-        public int in_reply_to_status_id;
-        public string in_reply_to_status_id_str;
-        public int in_reply_to_user_id;
-        public string in_reply_to_user_id_str;
-        public bool retweeted;
-        public int retweet_count;
-        public bool favorited;
-        public int favorite_count;
-        public bool is_quote_status;
-        public int quoted_status_id;
-        public string quoted_status_id_str;
-
-        public string lang;
-
-        public Entities entities;
-        public Extended_Entities extended_entities;
-    }
-
-    [Serializable]
-    public class TweetObjectWithUser: TweetObject
-    {
-        public TweetUser user;
-    }
-
-    [Serializable]
-    public class Tweet: TweetObjectWithUser
-    {
-        public TweetObjectWithUser retweeted_status;
-    }
-
-	[Serializable]
-	public class TweetUser {
-		public long id;
-		public string id_str;
-		public string name;
-		public string screen_name;
-		public string location;
-		public string description;
-		public string url;
-		public bool verified;
-
-		public string profile_text_color;
-		public bool profile_user_background_image;
-		public string profile_background_image_url;
-		public string profile_background_image_url_https;
-		public string profile_background_color;
-		public string profile_banner_url;
-		public bool profile_background_tile;
-		public string profile_image_url;
-
-		public int statuses_count;
-		public int friends_count;
-		public int followers_count;
-		public int favourites_count;
-		public bool following;
-		public bool follow_request_sent;
-		public int listed_count;
-
-		public TweetObject status;
-
-	}
-
-    [Serializable]
-    public class UploadMedia
-    {
-        public long media_id;
-        public string media_id_string;
-        public int size;
-        public int expires_after_secs;
-        public UploadMediaImage image;
-    }
-
-    [Serializable]
-    public class UploadMediaImage
-    {
-        public string image_type;
-        public int w;
-        public int h;
-    }
-
-    [Serializable]
-    public class List
-    {
-        public string slug;
-        public string name;
-        public string created_at;
-        public string uri;
-        public int subscriber_count;
-        public long id;
-        public string id_str;
-        public int member_count;
-        public string mode;
-        public string full_name;
-        public string description;
-        public TweetUser user;
-        public bool following;
-    }
-
-    #endregion
-
 
     # region Entities Class
