+++ scripts/twitterRequest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using Twity.Helpers;
 using UnityEngine;
 using UnityEngine.Networking;
